editoast: move all openapi logic into views::openapi
